Break down the server `Main` file
This commit breaks down the `Main` file to multiple different files each represents a specific part of the backend (e.g., model, route, etc...)
